// src/shared/types.ts
export enum SelectedPage {
    Home=  'home',
    Benefits= 'benefits',
    OurClasses = 'ourclasses',
    ContactUs = 'contactus'
}

export type BenefitType = {
    icon: JSX.Element
    title: string
    description: string
}

export type ClassType = {
    title: string,
    description?: string
    image: string
}